requirements.txt
setproctitle
aiohttp
msgpack-python